@@ -11,10 +11,10 @@ description: Evgeshayoga Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.8.0+61
+version: 1.8.1+62
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,7 +28,7 @@ dependencies:
   connectivity: ^3.0.2
   chewie: ^1.3.0
   firebase_auth: ^3.0.0
-  firebase_database: ^8.0.0
+  firebase_database: ^9.0.0
   firebase_crashlytics: ^2.0.0
   http: ^0.13.0
   url_launcher: ^6.0.2
