@@ -1,26 +1,28 @@
 import 'package:firebase_database/firebase_database.dart';
 
 class User {
-  int userId;
-  String userName;
-  String userEmail;
-  String password;
-  String phoneNumber;
+  late int userId;
+  late String userName;
+  late String userEmail;
+  late String password;
+  late String phoneNumber;
   UserPurchases? purchases;
 
   User(this.userName, this.userEmail, this.password, this.phoneNumber,
       {required this.userId, this.purchases});
 
-  User.fromSnapshot(DataSnapshot snapshot)
-      : userId = snapshot.value["id"],
-        userName = snapshot.value["name"],
-        userEmail = snapshot.value["email"],
-        phoneNumber = snapshot.value["phone"],
-        purchases = new UserPurchases(snapshot.value["purchases"]),
-        password = "";
+  User.fromSnapshot(DataSnapshot snapshot) {
+    var data = new Map<String, dynamic>.from(snapshot.value as dynamic);
+    userId = data["id"];
+    userName = data["name"];
+    userEmail = data["email"];
+    phoneNumber = data["phone"];
+    purchases = new UserPurchases(data["purchases"]);
+    password = "";
+  }
 
   UserPurchases getPurchases() {
-    return purchases != null ? purchases! : new UserPurchases({});
+    return purchases != null ? purchases! : new UserPurchases(null);
   }
 }
 
